+++ aws/endpoints/defaults.go
@@ -2594,7 +2594,8 @@ var awscnPartition = partition{
 		"sms": service{
 
 			Endpoints: endpoints{
-				"cn-north-1": endpoint{},
+				"cn-north-1":     endpoint{},
+				"cn-northwest-1": endpoint{},
 			},
 		},
 		"snowball": service{